requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
apscheduler==3.10.4
requests==2.31.0
PyYAML==6.0.1
pytz==2024.1
plyer==2.1.0
pydantic==2.6.1
